azure_guardian/requirements.txt
azure-functions
GitPython
PyYAML
scikit-learn
joblib

